Fix incorrect array type in example code.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@1757 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ users/base/user_modules.f90
@@ -42,8 +42,8 @@ module user_module
 
   ! Example: allocatable user arrays
 
-  integer, dimension(:), allocatable :: iwork
-  integer, dimension(:,:), allocatable :: rwork
+  integer,          dimension(:), allocatable :: iwork
+  double precision, dimension(:,:), allocatable :: rwork
 
 contains
